requirements.txt
flask
google-generativeai
openai